requirements.txt
Flask==1.1.1
gunicorn==19.9.0
line-bot-sdk==1.14.0
mysql-connector-python==8.0.26
numpy==1.19.5
pandas==1.1.5
requests==2.26.0